+++ modules_s/avp_db/avp_db.c
@@ -335,7 +335,7 @@ static int load_avp(struct sip_msg* msg, char* attr, char* _dummy)
 
 static int load_avp_fixup(void** param, int param_no)
 {
-	int id = 0;
+	long id = 0;
 
 	if (param_no == 1) {
 		if (!strcasecmp(*param, "caller_uuid")) {
